chore: move MouseMode enum to a mouse subfolder
