+++ lib/src/core/mouse/mode.dart
