+++ lib/src/utils/debugger.dart
@@ -1,6 +1,6 @@
 import 'package:xterm/src/core/escape/handler.dart';
 import 'package:xterm/src/core/escape/parser.dart';
-import 'package:xterm/src/core/mouse.dart';
+import 'package:xterm/src/core/mouse/mode.dart';
 import 'package:xterm/src/utils/observable.dart';
 
 class TerminalCommand {
